A last step to complete project 01
At this stage, it becomes a bit boring to continue.
examples are longer, our brain needs to keep track of more information,
we can not get the thrill of completion easily.
This is the difference between a newbie and a master, once we went
through this, all concepts will be iron on our brain, this bordness and
pain will give you great revenue.